move call_brain_api
